Create README.md
The grocery market in India offers affordability and accessibility to a diverse range of products, benefiting consumers. It generates employment opportunities, supports local farmers and suppliers, and contributes to the country's economic growth. The market ensures food security, convenience, and adheres to health and safety standards, providing consumers with quality food options.

https://7by24info.com/